+++ lib/ecto/query/builder/group_by.ex
@@ -14,11 +14,49 @@ defmodule Ecto.Query.Builder.GroupBy do
        %{}}
   """
   @spec escape(Macro.t, Keyword.t, Macro.Env.t) :: Macro.t
+  def escape({:^, _, [expr]}, _vars, _env) do
+    {quote(do: Ecto.Query.Builder.GroupBy.group_by!(unquote(expr))), %{}}
+  end
+
   def escape(expr, vars, env) do
     expr
     |> List.wrap
-    |> Builder.escape(:any, %{}, vars, env)
+    |> Enum.map_reduce(%{}, &do_escape(&1, &2, vars, env))
   end
+
+  defp do_escape({:^, _, [expr]}, params, _vars, _env) do
+    {quote(do: Ecto.Query.Builder.GroupBy.field!(unquote(expr))), params}
+  end
+
+  defp do_escape(field, params, _vars, _env) when is_atom(field) do
+    {Macro.escape(to_field(field)), params}
+  end
+
+  defp do_escape(expr, params, vars, env) do
+    Builder.escape(expr, :any, params, vars, env)
+  end
+
+  @doc """
+  Called at runtime to verify a field.
+  """
+  def field!(field) when is_atom(field),
+    do: to_field(field)
+  def field!(other),
+    do: Builder.error!("expected a field as an atom in `group_by`, got: `#{inspect other}`")
+
+  @doc """
+  Called at runtime to verify group_by.
+  """
+  def group_by!(group_by) do
+    Enum.map List.wrap(group_by), fn
+      field when is_atom(field) ->
+        to_field(field)
+      _ ->
+        Builder.error!("expected a list of fields in `group_by`, got: `#{inspect group_by}`")
+    end
+  end
+
+  defp to_field(field), do: {{:., [], [{:&, [], [0]}, field]}, [], []}
 
   @doc """
   Builds a quoted expression.